a2a-testclient/requirements.txt
a2a-sdk>=0.3.9
google-auth-oauthlib
requests
